send log file as attachment
gmail clips the log forcing you to login to the server to look a it.
Send it as an attachment. This commit also does not create the SMTP
server unless we are sending emails. This removes the round trip to the
server to authenticate. I've also added a spot to create local emails as
emls for testing.