Filter out empty geometries in update cursor
This prevents the update cursor for erroring with the following error:
Traceback (most recent call last):
  File "c:\python27\arcgis10.3\lib\site-packages\forklift\core.py",
line 162, in _move_data
    icursor.insertRow(row)
RuntimeError: The field is not nullable. [Shape]

This was being caused by a point feature that had no coordinates in its
shape field. I was unsuccessful in trying to use queries to filter out
these problem records so I fell back onto the if statement.