feat: add fixture for auto-zipping/unzipping of test fgdbs This commit defines a pytest fixture named `test_gdb`. To use, define a `test_gdb` parameter in your test method.
Before the test is run, the fixture looks for an existing fgdb in the path: `data/<test_file>/<test_name>.gdb`. If it finds it, it copies it to a temp directory and returns the path to the test. If it doesn't find an existing fgdb, it looks for a zip in the same type of path: `data/<test_file>/<test_name>.zip`. If it find a zip, it extracts it to a temp folder and returns the path to the test.
After the test has run, if it was successful and there was a fgdb in the data folder, it zips it up and deletes the original fgdb.

The temp directories are unique to each test run are cleaned up after a few cycles.

Caveats
This only works on stand-alone test functions, not class methods. This is because class methods only support "autouse" fixtures.